📜 update explicit dependency declarations
When I first tired to build this app it failed, because it couldn't
download a dependency. This didn't hit any other developers yet, because
gradle caches dependencies (on linux in ~/.gradle/cache). So what I did
for this change is to delete my cache and updated the dependency listing
accordingly, until I could build the app from an empty cache.